Create a self-signed cert before provision
There's a separate update on forumone-puppet that adds HTTPS by default
for Apache and NGINX. NGINX (and puppet in general) lacks the ability to
generate certs for itself, so we have to do it here. This step is
redundant when hosting on Apache, or using the box without a web server.