Update drupal.yaml
Per README for features, 3000 for max input is recommended if we are exporting a lot to features. We do on a regular basis. So let's bump to that recommendation:

<blockquote>If you plan on creating or working with very large features (greater than 1000
items), you may need to increase PHP's max_input_vars configuration directive.
For example, adding the following line to your .htaccess file will increase the
max_input_vars directive to 3000:
</blockquote>